+++ node/src/Transaction.ts
@@ -5,7 +5,8 @@
 import {
     BaseClient, // eslint-disable-line @typescript-eslint/no-unused-vars
     GlideString,
-    ReadFrom, // eslint-disable-line @typescript-eslint/no-unused-vars
+    HashDataType,
+    convertFieldsAndValuesForHset,
 } from "./BaseClient";
 
 import {
@@ -806,13 +807,18 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      * @see {@link https://valkey.io/commands/hset/|valkey.io} for details.
      *
      * @param key - The key of the hash.
-     * @param fieldValueMap - A field-value map consisting of fields and their corresponding values
+     * @param fieldValueList - A list of field names and their values.
      * to be set in the hash stored at the specified key.
      *
      * Command Response - The number of fields that were added.
      */
-    public hset(key: GlideString, fieldValueMap: Record<string, string>): T {
-        return this.addAndReturn(createHSet(key, fieldValueMap));
+    public hset(
+        key: GlideString,
+        fieldsAndValues: HashDataType | Record<string, GlideString>,
+    ): T {
+        return this.addAndReturn(
+            createHSet(key, convertFieldsAndValuesForHset(fieldsAndValues)),
+        );
     }
 
     /**
@@ -2534,8 +2540,8 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      * Command Response - The id of the added entry, or `null` if `options.makeStream` is set to `false` and no stream with the matching `key` exists.
      */
     public xadd(
-        key: string,
-        values: [string, string][],
+        key: GlideString,
+        values: [GlideString, GlideString][],
         options?: StreamAddOptions,
     ): T {
         return this.addAndReturn(createXAdd(key, values, options));
@@ -2552,7 +2558,7 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      * Command Response - The number of entries removed from the stream. This number may be less than the number of entries in
      *      `ids`, if the specified `ids` don't exist in the stream.
      */
-    public xdel(key: string, ids: string[]): T {
+    public xdel(key: GlideString, ids: GlideString[]): T {
         return this.addAndReturn(createXDel(key, ids));
     }
 
@@ -2790,11 +2796,11 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      * Command Response - A `Record` of message entries that are claimed by the consumer.
      */
     public xclaim(
-        key: string,
-        group: string,
-        consumer: string,
+        key: GlideString,
+        group: GlideString,
+        consumer: GlideString,
         minIdleTime: number,
-        ids: string[],
+        ids: GlideString[],
         options?: StreamClaimOptions,
     ): T {
         return this.addAndReturn(
@@ -2854,11 +2860,11 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      *     These IDs are deleted from the Pending Entries List.
      */
     public xautoclaim(
-        key: string,
-        group: string,
-        consumer: string,
+        key: GlideString,
+        group: GlideString,
+        consumer: GlideString,
         minIdleTime: number,
-        start: string,
+        start: GlideString,
         count?: number,
     ): T {
         return this.addAndReturn(
@@ -2924,9 +2930,9 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      * Command Response - `"OK"`.
      */
     public xgroupCreate(
-        key: string,
-        groupName: string,
-        id: string,
+        key: GlideString,
+        groupName: GlideString,
+        id: GlideString,
         options?: StreamGroupOptions,
     ): T {
         return this.addAndReturn(
@@ -2944,7 +2950,7 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      *
      * Command Response - `true` if the consumer group is destroyed. Otherwise, `false`.
      */
-    public xgroupDestroy(key: string, groupName: string): T {
+    public xgroupDestroy(key: GlideString, groupName: GlideString): T {
         return this.addAndReturn(createXGroupDestroy(key, groupName));
     }
 
@@ -2960,9 +2966,9 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      * Command Response - `true` if the consumer is created. Otherwise, returns `false`.
      */
     public xgroupCreateConsumer(
-        key: string,
-        groupName: string,
-        consumerName: string,
+        key: GlideString,
+        groupName: GlideString,
+        consumerName: GlideString,
     ): T {
         return this.addAndReturn(
             createXGroupCreateConsumer(key, groupName, consumerName),
@@ -2981,9 +2987,9 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      * Command Response - The number of pending messages the `consumer` had before it was deleted.
      */
     public xgroupDelConsumer(
-        key: string,
-        groupName: string,
-        consumerName: string,
+        key: GlideString,
+        groupName: GlideString,
+        consumerName: GlideString,
     ): T {
         return this.addAndReturn(
             createXGroupDelConsumer(key, groupName, consumerName),
@@ -3002,7 +3008,7 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      *
      * Command Response - The number of messages that were successfully acknowledged.
      */
-    public xack(key: string, group: string, ids: string[]): T {
+    public xack(key: GlideString, group: GlideString, ids: GlideString[]): T {
         return this.addAndReturn(createXAck(key, group, ids));
     }
 
@@ -3020,9 +3026,9 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      * Command Response - `"OK"`.
      */
     public xgroupSetId(
-        key: string,
-        groupName: string,
-        id: string,
+        key: GlideString,
+        groupName: GlideString,
+        id: GlideString,
         entriesRead?: number,
     ): T {
         return this.addAndReturn(